refactor: 숫자 포맷팅 함수 별도의 파일로 분리
